+++ crates/client/src/client.rs
@@ -386,7 +386,7 @@ impl FuelClient {
         &self,
         txs: &[Transaction],
     ) -> io::Result<Vec<TransactionExecutionStatus>> {
-        self.dry_run_opt(txs, None).await
+        self.dry_run_opt(txs, None, None).await
     }
 
     /// Dry run with options to override the node behavior
@@ -395,6 +395,7 @@ impl FuelClient {
         txs: &[Transaction],
         // Disable utxo input checks (exists, unspent, and valid signature)
         utxo_validation: Option<bool>,
+        gas_price: Option<u64>,
     ) -> io::Result<Vec<TransactionExecutionStatus>> {
         let txs = txs
             .iter()
@@ -404,6 +405,7 @@ impl FuelClient {
             schema::tx::DryRun::build(DryRunArg {
                 txs,
                 utxo_validation,
+                gas_price: gas_price.map(|gp| gp.into()),
             });
         let tx_statuses = self.query(query).await.map(|r| r.dry_run)?;
         tx_statuses