+++ crates/client/src/client/schema/tx.rs
@@ -370,6 +370,7 @@ pub struct EstimatePredicates {
 pub struct DryRunArg {
     pub txs: Vec<HexString>,
     pub utxo_validation: Option<bool>,
+    pub gas_price: Option<U64>,
 }
 
 #[derive(cynic::QueryFragment, Clone, Debug)]
@@ -379,7 +380,7 @@ pub struct DryRunArg {
     variables = "DryRunArg"
 )]
 pub struct DryRun {
-    #[arguments(txs: $txs, utxoValidation: $utxo_validation)]
+    #[arguments(txs: $txs, utxoValidation: $utxo_validation, gasPrice: $gas_price)]
     pub dry_run: Vec<DryRunTransactionExecutionStatus>,
 }
 
@@ -467,7 +468,8 @@ pub mod tests {
         let tx = fuel_tx::Transaction::default_test_tx();
         let query = DryRun::build(DryRunArg {
             txs: vec![HexString(Bytes(tx.to_bytes()))],
-            utxo_validation: None,
+            utxo_validation: Some(true),
+            gas_price: Some(123u64.into()),
         });
         insta::assert_snapshot!(query.query)
     }